Added a solution to exercise 4.6

+++ 4_6/src/main.c
@@ -19,8 +19,8 @@ int main() {
   char* input_one = malloc(sizeof(char));
   char* input_two = malloc(sizeof(char));
 
-  int user_input1 = 0;
-  int user_input2 = 0;
+  long user_input1 = 0;
+  long user_input2 = 0;
 
   while (1) {
     // User input
@@ -36,14 +36,13 @@ int main() {
     user_input2 = strtol(input_two, &endptr2, sizeof(input_two));
 
     // Bruger FindGcd funktionen til at finde GCD
-    int gcd = FindGcd(user_input1, user_input2);
+    unsigned long gcd = FindGcd(user_input1, user_input2);
     if (user_input1 <= 0 || user_input2 <= 0) {
       // Error handling
       printf("Error: One or more numbers are not positive integers\n");
       break;
-    } printf("Greatest common divisor between %d and %d is %d\n", user_input1,
-           user_input2, gcd);
-
+    } printf("Greatest common divisor between %lu and %lu is %lu\n",
+           user_input1,user_input2, gcd);
   }
 
   free(input_one);
